Updated "PaletteManager::setPaletteSplitPositionY" to support parameters larger than u8

+++ Oxygen/oxygenengine/source/oxygen/rendering/parts/palette/PaletteManager.cpp
@@ -122,7 +122,7 @@ void PaletteManager::writePaletteEntryPacked(int paletteIndex, uint16 colorIndex
 	mMainPalette[1].setPaletteEntryPacked(colorIndex, color, packedColor);
 }
 
-void PaletteManager::setPaletteSplitPositionY(uint8 py)
+void PaletteManager::setPaletteSplitPositionY(int py)
 {
 	mSplitPositionY = py;
 }
